Do not throw exception in FileSystemBackupStorage when input path=null (#25)
Current FileSystemBackupStorage implementation has some inconsistencies with existing open source implementation for HdfsBackupStorage. FileSystemBackupStorage throws a backup exception when the input File path is null that saying the provided path is invalid, while HdfsBackupStorage does not throw exceptions. These inconsistencies result to the need of special handling in test cases. To ensure that all the implementations of BackupStorageProvider have same semantics, this commit will remove the exceptions thrown by FileSystemBackupStorage in the above described situations, and return empty list/array as how it is handled in HdfsBackupStorage.