Fix issues when remove invalid temporary backup files (#28)
An IOException from FileSystemBackupStorage in copyToBackupStorage() is found during our testing. The root cause is that currently when cleanupInvalidFiles() method is called, all the temporary backup files (both transaction logs and snapshots) will be deleted. And this caused an undesired situation where the snapshot backup thread deletes temporary transaction log backup files before they can be saved as formal backup files.
This commit fixes this bug from two aspects:

Add a read write lock to coordinate the temporary files' read and write access from multiple threads in FileSystemBackupStorage
Modify the way to log an exception in BackupManager to allow detailed exception stack trace to show in the logs, for easier debugging in the future