Two minor improvements on BackupManager and FileSystemBackupStorage (#33)
This commit includes two minor improvements:
Removed duplicated calls of initialize() in BackupManager. Currently initialize() is called twice: once inside BackupManager constructor, and once inside start().
Add a check for backup storage path existence in FileSystemBackupStorage constructor. This is to ensure we fail it early if there is something wrong with the setup of backup storage, instead of writing backup files to unexpected places.