Add timetable support for ZooKeeper backup and restore (#32)
This commit adds the support for timetable (data model and backup thereof). The backup timetable is to be used to look up the closest zxid given the timestamp to enable point-in-time restoration. More specifically, this commit includes:
1. definition of timetable metadata, timetable backup process, timetable recorder child thread
2. moves BackupProcess thread out so that it is no longer an inner class
3. TimetableUtil class that provides static methods for timetable-zxid lookup
4. unit and integration tests: BackupManagerTest::testTimetableBackup() and TimetableUtilTest
5. fix an unstable test: BackupManagerTest::testBackupManager()

There will be follow-up PRs for 1) timetable backup status check and 2) timetable backup metrics.