Add timetable to backup status file (#38)
The backup status file stored in statusDir for ZooKeeper backup is used to coordinate between snapshot and TxLog backup processes. It is used in case of failure or restart so that the respective threads can pick up where they left off before restarting. This commit adds timetable to the status file so that in case of restart, the timetable backup process will also pick up where it left off and retry any temporary timetable backup files on the backup server.