Add timetable backup metrics (#40)
Timetable backup is supported as part of ZooRestore in order to enable users to restore to a point in time using a timestamp value. Since the timetable backup process is similar to snapshot and transaction log backups, we need metrics to monitor whether the backup of timetable metadata is working correctly. This commit adds the necessary metrics.

test3_TimetableBackupMBean has been added to the existing BackupBeanTest class.