Add restore command to ZK CLI (#39)
This commit adds a CLI command to ZooKeeperMain, and it will allow zk admin to run commands on a machine to restore the backed-up zk snapshot and transaction log files up until a specified zxid or timestamp from backup storage to a local path. This enables the offline restoration of a zk server node to a desired time/zxid.